requirements.txt
streamlit
langchain-openai
langchain-community
langchain-chroma
pypdf
python-dotenv